+++ staircase/test_data.py
@@ -66,21 +66,30 @@ def make_test_data(dates=True, positive_only=True, groups=(), seed=None):
     return pd.concat(dfs)
 
 
-def _make_test_data(dates=True, positive_only=True, seed=None):
-    if seed is None:
-        rng = np.random.default_rng()
-    else:
-        rng = np.random.default_rng(seed)
+def _get_random_functions(seed=None):
+    major, minor, *_ = np.__version__.split(".")
 
-    n_intervals = rng.integers(100, 1000)
-    smallest_interval = rng.integers(1, 6)
-    largest_interval = rng.integers(15, 20)
+    if f"{major}.{minor}" in ("1.14", "1.15", "1.16"):
+        rs = np.random.RandomState(seed)
+        return rs.randint, rs.uniform
+
+    rng = np.random.default_rng(seed)
+    return rng.integers, rng.random
+
+
+def _make_test_data(dates=True, positive_only=True, seed=None):
+
+    random_integers, random_floats = _get_random_functions(seed)
+
+    n_intervals = random_integers(100, 1000)
+    smallest_interval = random_integers(1, 6)
+    largest_interval = random_integers(15, 20)
     interval_lengths = (
-        rng.random(size=n_intervals) * (largest_interval - smallest_interval)
+        random_floats(size=n_intervals) * (largest_interval - smallest_interval)
         + smallest_interval
     )
 
-    start = rng.random(size=n_intervals) * 120 - 20
+    start = random_floats(size=n_intervals) * 120 - 20
     end = start + interval_lengths
 
     if dates:
@@ -91,11 +100,17 @@ def _make_test_data(dates=True, positive_only=True, seed=None):
             end * 365 / 100 * pd.Timedelta(1, "day")
         )
 
-    value = rng.choice(range(-5, 6), n_intervals)
+    value = random_integers(-5, 6, n_intervals)
     if positive_only:
         value = value + 6
 
-    data = pd.DataFrame({"start": start, "end": end, "value": value,})
+    data = pd.DataFrame(
+        {
+            "start": start,
+            "end": end,
+            "value": value,
+        }
+    )
 
     if dates:
         data["start"] = data["start"].dt.floor("min")