+++ docs/release_notes/changelog.rst
@@ -9,6 +9,7 @@ UNRELEASED
 
 
 - dropped support for Python 3.6
+- bugfix for :func:`staircase.make_test_data` which failed with numpy < 1.17 (#GH80)
 
 Please list new changes above this comment